requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.0.9
    # via kombu
asgiref==3.4.1
    # via django
asttokens==2.0.5
    # via stack-data
backcall==0.2.0
    # via ipython
billiard==3.6.4.0
    # via celery
black==21.12b0
    # via ipython
brotli==1.0.9
    # via fonttools
brotlipy==0.7.0
    # via -r requirements.in
celery[redis]==5.2.3
    # via
    #   -r requirements.in
    #   celery-redbeat
celery-redbeat==2.0.0
    # via -r requirements.in
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via
    #   brotlipy
    #   weasyprint
charset-normalizer==2.0.10
    # via requests
click==8.0.3
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cssselect2==0.4.1
    # via weasyprint
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via odfpy
deprecated==1.2.13
    # via redis
diff-match-patch==20200713
    # via django-import-export
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.9
    # via
    #   -r requirements.in
    #   django-admin-sortable
    #   django-extensions
    #   django-filter
    #   django-import-export
    #   django-js-reverse
    #   django-log-request-id
    #   django-model-utils
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-admin-sortable==2.3
    # via -r requirements.in
django-debreach==2.1.0
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-import-export==2.7.1
    # via -r requirements.in
django-js-reverse==0.9.1
    # via -r requirements.in
django-log-request-id==2.0.0
    # via -r requirements.in
django-model-utils==4.2.0
    # via -r requirements.in
django-webpack-loader==1.4.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
    #   rest-framework-generic-relations
djangorestframework-simplejwt==5.0.0
    # via -r requirements.in
drf-writable-nested==0.6.3
    # via -r requirements.in
drf-yasg==1.20.0
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
executing==0.8.2
    # via stack-data
fonttools[woff]==4.28.5
    # via weasyprint
gunicorn==20.1.0
    # via -r requirements.in
html5lib==1.1
    # via weasyprint
idna==3.3
    # via requests
importlib-metadata==4.10.1
    # via markdown
inflection==0.5.1
    # via drf-yasg
ipython==8.0.1
    # via -r requirements.in
itypes==1.2.0
    # via coreapi
jedi==0.18.1
    # via ipython
jinja2==3.0.3
    # via coreschema
kombu==5.2.3
    # via celery
markdown==3.3.6
    # via -r requirements.in
markuppy==1.14
    # via tablib
markupsafe==2.0.1
    # via jinja2
matplotlib-inline==0.1.3
    # via ipython
mypy-extensions==0.4.3
    # via black
numpy==1.22.1
    # via pandas
odfpy==1.4.1
    # via tablib
openpyxl==3.0.9
    # via tablib
packaging==21.3
    # via
    #   drf-yasg
    #   redis
pandas==1.3.5
    # via -r requirements.in
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.0.0
    # via weasyprint
platformdirs==2.4.1
    # via black
prompt-toolkit==3.0.24
    # via
    #   click-repl
    #   ipython
psutil==5.9.0
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.1
    # via stack-data
pycparser==2.21
    # via cffi
pydyf==0.1.2
    # via weasyprint
pygments==2.11.2
    # via ipython
pyjwt==2.1.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
pyparsing==3.0.6
    # via packaging
pyphen==0.12.0
    # via weasyprint
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   celery-redbeat
    #   pandas
python-decouple==3.5
    # via -r requirements.in
pytz==2021.3
    # via
    #   celery
    #   django
    #   djangorestframework
    #   pandas
pyyaml==6.0
    # via tablib
redis==4.1.1
    # via
    #   celery
    #   celery-redbeat
requests==2.27.1
    # via coreapi
rest-framework-generic-relations==2.1.0
    # via -r requirements.in
ruamel.yaml==0.17.20
    # via drf-yasg
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
sentry-sdk==1.5.2
    # via -r requirements.in
shortuuid==1.0.8
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   click-repl
    #   html5lib
    #   python-dateutil
sqlparse==0.4.2
    # via django
stack-data==0.1.4
    # via ipython
tablib[html,ods,xls,xlsx,yaml]==3.1.0
    # via django-import-export
tenacity==8.0.1
    # via celery-redbeat
tinycss2==1.1.1
    # via
    #   cssselect2
    #   weasyprint
tomli==1.2.3
    # via black
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.0.1
    # via black
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.8
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
weasyprint==53.4
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
whitenoise==5.3.0
    # via -r requirements.in
wrapt==1.13.3
    # via deprecated
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
zipp==3.7.0
    # via importlib-metadata
zopfli==0.1.9
    # via fonttools

# The following packages are considered to be unsafe in a requirements file:
# setuptools
